Remove select checkboxes from user lists

+++ app/helpers/admin/candidates_helper.rb
@@ -3,7 +3,6 @@ module Admin::CandidatesHelper
     returning [] do |arr|
       arr << %(<table class="tbl-standard">)
       arr << %(<tr>)
-      arr << %(<th>#{check_box_tag('select-all')}</th>)
       arr << %(<th>First name</th>)
       arr << %(<th>Last name</th>)
       arr << %(<th>Email</th>)
@@ -21,7 +20,6 @@ private
     returning [] do |arr|
       users.each_with_index do |row, row_index|
         arr << %(<tr class="#{row_index % 2 == 0 ? 'even': 'odd'}">)
-        arr << %(<td>#{check_box_tag(row.id)}</td>)
         arr << %(<td>#{row.firstname}</td>)
         arr << %(<td>#{row.lastname}</td>)
         arr << %(<td>#{row.email}</td>)
@@ -29,7 +27,7 @@ private
         arr << %(<td>#{cv_link(row)}</td>)
         arr << %(<td>#{image_link_to('icons/ico-view.png', 'view', admin_candidate_path(row.id))})
         arr << %(  #{image_link_to('icons/ico-edit.png','edit', edit_admin_candidate_path(row.id))})
-        arr << %(  #{image_link_to('icons/ico-del.png', 'delete', { :action => :destroy, :id => row }, nil, :method => :delete)}</td>)
+        arr << %(  #{image_link_to('icons/ico-del.png', 'delete', { :action => :destroy, :id => row }, nil, { :method => :delete, :confirm => "Are you sure?" } )}</td>)
         arr << %(</tr>)
       end
     end
